+++ release_notes.md
@@ -10,6 +10,7 @@
     - added support for `CTkTextBoxEx` targets
   - added global `on_edit_begin` and `on_edit_end` hooks, e.g. to show/hide keyboard when editing begins/ends
   - keyboard now detects disabled entries and doesn't start editing them when clicked
+  - keymap can define hover color
 - `el.datastore.SavableModel`
   - added ClassVar `model_dump_default_options` to  to specify default options used for saving files (like indentation).
 - `el.widgets.ctkex.CTkEntryEx`
@@ -18,8 +19,8 @@
   - added `dark_when_disabled` option: Makes button fg the hover_color when it is in disabled state
 - `el.widgets.listbox.CTkListBox`:
   - added methods `select_single_by_index`, `select_single_by_id`, `deselect_all`
-- `el.numbers`
-  - added `clamp` function to clamp value between a minimum and maximum
+- `el.observable.Observable` and `el.observable.ComposedObservable`
+  - `force_notify()` now propagates force updates recursively to all descendant observables, composed observables and observers, mitigating issues where forcing an update after value mutation (e.g. appending to a list) only propagated to the first layer (direct observers). This was achieved by adding the `force_recursive` kwarg throughout the entire value update call path, optionally including observer functions.
 
 ## New Features
 
@@ -31,3 +32,7 @@
   - ability to exclude corners from radius
   - ability to change background corner colors
   - ability to control width/height rounding
+- `el.numbers`
+  - added `clamp` function to clamp value between a minimum and maximum
+- `el.ctk_utils`
+  - added `apply_enabled` function to more easily set widget "state" from boolean value